new updated file

+++ database/migrations/2025_07_16_033008_add_hour_to_users_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up()
     {
     Schema::table('users', function (Blueprint $table) {
-        $table->integer('hour')->nullable();
+        $table->float('hour')->nullable();
     });
     }
